requirements.txt
pytest-playwright
pytest
playwright
faker
allure-pytest
pytest-dotenv